🧱 Front-end

    - ✨ Desativando o Servidor de Rotas.
    - ✨ Desativando a validação de Sessão.

🧱 Back-end

    - ✨ Desativando a validação de Sessão.
    - ✨ Implementando Response e Status em TipoProcesso.

🚧 Em análise e desenvolvimento

    - 🔍 Avaliar os processos de validação de rotas do servidor.
    - 🔍 Verificar a persistência do token como cookie salvo no lado do cliente.
    - 📝 Possível necessidade de requisitar ao Governo a lógica para validação de IE.
    
    - 📚 Estudo e implementação da tela multisserviço do TipoDocumentoEtapa (front-end).
    - 📚 Corrigir a lógica de validação do token no middleware pós-requisição (back-end).

+++ Desenvolvimento/BackEnd/API/SGED/Repositories/Entities/TipoProcessoRepository.cs
@@ -1,49 +1,52 @@
 ﻿using SGED.Context;
-using SGED.Models.Entities;
 using SGED.Repositories.Interfaces;
 using Microsoft.EntityFrameworkCore;
 using Npgsql;
+using SGED.Objects.Models.Entities;
 
 namespace SGED.Repositories.Entities;
 public class TipoProcessoRepository : ITipoProcessoRepository
 {
 
-	private readonly AppDBContext _dbContext;
+    private readonly AppDBContext _dbContext;
 
-	public TipoProcessoRepository(AppDBContext dbContext)
-	{
-		_dbContext = dbContext;
-	}
+    public TipoProcessoRepository(AppDBContext dbContext)
+    {
+        _dbContext = dbContext;
+    }
 
-	public async Task<IEnumerable<TipoProcesso>> GetAll()
-	{
-		return await _dbContext.TipoProcesso.ToListAsync();
-	}
+    public async Task<IEnumerable<TipoProcesso>> GetAll()
+    {
+        return await _dbContext.TipoProcesso.AsNoTracking().ToListAsync();
+    }
 
-	public async Task<TipoProcesso> GetById(int id)
-	{
-		return await _dbContext.TipoProcesso.Where(objeto => objeto.Id == id).FirstOrDefaultAsync();
-	}
+    public async Task<TipoProcesso> GetById(int id)
+    {
+        return await _dbContext.TipoProcesso.AsNoTracking().FirstOrDefaultAsync(objeto => objeto.Id == id);
+    }
 
-	public async Task<TipoProcesso> Create(TipoProcesso tipoProcesso)
-	{
-		_dbContext.TipoProcesso.Add(tipoProcesso);
-		await _dbContext.SaveChangesAsync();
-		return tipoProcesso;
-	}
+    public async Task<TipoProcesso> Create(TipoProcesso tipoProcesso)
+    {
+        _dbContext.TipoProcesso.Add(tipoProcesso);
+        await _dbContext.SaveChangesAsync();
+        return tipoProcesso;
+    }
 
-	public async Task<TipoProcesso> Update(TipoProcesso tipoProcesso)
-	{
-		_dbContext.Entry(tipoProcesso).State = EntityState.Modified;
-		await _dbContext.SaveChangesAsync();
-		return tipoProcesso;
-	}
+    public async Task<TipoProcesso> Update(TipoProcesso tipoProcesso)
+    {
+        _dbContext.Entry(tipoProcesso).State = EntityState.Modified;
+        await _dbContext.SaveChangesAsync();
+        return tipoProcesso;
+    }
 
-	public async Task<TipoProcesso> Delete(int id)
-	{
-		var tipoProcesso = await GetById(id);
-		_dbContext.TipoProcesso.Remove(tipoProcesso);
-		await _dbContext.SaveChangesAsync();
-		return tipoProcesso;
-	}
+    public async Task<TipoProcesso> Delete(int id)
+    {
+        var tipoProcesso = await GetById(id);
+        if (tipoProcesso != null)
+        {
+            _dbContext.TipoProcesso.Remove(tipoProcesso);
+            await _dbContext.SaveChangesAsync();
+        }
+        return tipoProcesso;
+    }
 }
